+++ lib/db/stats.js
@@ -1,4 +1,4 @@
-import {findByDay, findByField} from './driver.js'
+import {findByDay, findByField} from './mongo.js'
 
 
 export async function readSiteStats(host) {
