+++ lib/server/counter.js
@@ -1,5 +1,5 @@
 import {promises as fs} from 'fs'
-import {insertOne} from '../db/driver.js'
+import {insertOne} from '../db/mongo.js'
 import geoip from 'geoip-lite'
 import DeviceDetector from 'node-device-detector'
